+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_responses.hpp
@@ -87,7 +87,10 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @return A reference to an #Azure::Core::Http::RawResponse.
      * @note Does not give up ownership of the RawResponse.
      */
-    Azure::Core::Http::RawResponse const& GetRawResponse() const override { return *m_rawResponse; }
+    Azure::Core::Http::RawResponse const& GetRawResponseInternal() const override
+    {
+      return *m_rawResponse;
+    }
 
     StartCopyBlobOperation() = default;